backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
python-jose[cryptography]
passlib[pbkdf2_sha256]
python-multipart
python-dotenv


langchain-community
langchain-text-splitters
langchain-huggingface
pypdf
faiss-cpu
sentence-transformers
joblib
numpy

langchain-google-genai